refactor(imagegen): make imagegen a leaf package owned by the OpenAI client
Move internal/client/imagegen -> internal/imagegen and invert the
dependency: client.OpenAIClient now imports imagegen, instead of the
server reaching into imagegen and injecting a delegate back.

OpenAIClient.ImagesGenerate dispatches DashScope / MiniMax to their
native imagegen adapters and serves every OpenAI-compatible provider
through the SDK directly. Codex keeps its own CodexClient.ImagesGenerate
Responses-API path. The server handler and forwarding layer no longer
know imagegen exists — they just call ImagesGenerate on the client
wrapper, so the external surface stays "OpenAI client + Codex client".

Drops the NewDelegate / delegateClient seam, the ErrDelegateRequired and
ErrResponsesAPIRequired sentinels, and ForwardImageGeneration, all of
which only existed to work around the previous dependency direction.
