+++ internal/imagegen/minimax.go
